# openapi/ziji-v1.yaml
...
openapi: 3.1.0
info:
  title: 资迹 Ziji API
  version: 0.1.0
  description: V1 正式 API 契约，对应 PRD V0.1、RTM V0.1 和《API 契约》V0.1；96 个 HTTP operationId 为正式实施基线。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: 当前主版本
security:
  - bearerAuth: []
...
paths:
  /auth/registration-challenges:
    post:
      tags: [Auth]
      operationId: createRegistrationChallenge
      summary: 发送注册邮箱验证码
      security: []
      x-requirement-ids: [AUTH-001, SEC-002]
      requestBody: { $ref: '#/components/requestBodies/EmailChallengeRequest' }
      responses:
        '202': { $ref: '#/components/responses/ChallengeAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: 验证邮箱验证码并注册
      security: []
      x-requirement-ids: [AUTH-001]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } }
      responses:
        '201': { $ref: '#/components/responses/UserCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /auth/web/sessions:
    post:
      tags: [Auth]
      operationId: createWebSession
      summary: Web 邮箱密码登录
      security: []
      x-requirement-ids: [AUTH-002, AUTH-007]
      requestBody: { $ref: '#/components/requestBodies/LoginRequest' }
      responses:
        '201': { $ref: '#/components/responses/WebSessionCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/web/sessions/refresh:
    post:
      tags: [Auth]
      operationId: refreshWebSession
      summary: 使用 HttpOnly Cookie 轮换 Web 会话
      security:
        - webRefreshCookie: []
      x-requirement-ids: [AUTH-002, AUTH-007, SEC-001]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200': { $ref: '#/components/responses/WebSessionRefreshed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/mobile/sessions:
    post:
      tags: [Auth]
      operationId: createMobileSession
      summary: Mobile 邮箱密码登录
      security: []
      x-requirement-ids: [AUTH-002, AUTH-007]
      requestBody: { $ref: '#/components/requestBodies/LoginRequest' }
      responses:
        '201': { $ref: '#/components/responses/MobileSessionCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/mobile/sessions/refresh:
    post:
      tags: [Auth]
      operationId: refreshMobileSession
      summary: 轮换 Mobile 刷新 Token
      security: []
      x-requirement-ids: [AUTH-002, AUTH-007, SEC-001]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/MobileRefreshRequest' } }
      responses:
        '200': { $ref: '#/components/responses/MobileSessionRefreshed' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /auth/password-reset-challenges:
    post:
      tags: [Auth]
      operationId: createPasswordResetChallenge
      summary: 发送密码重置验证码
      security: []
      x-requirement-ids: [AUTH-003, SEC-002]
      requestBody: { $ref: '#/components/requestBodies/EmailChallengeRequest' }
      responses:
        '202': { $ref: '#/components/responses/ChallengeAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/password-reset:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: 重置密码并撤销既有会话
      security: []
      x-requirement-ids: [AUTH-003]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/PasswordResetRequest' } }
      responses:
        '204': { description: 密码已重置 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      summary: 查询当前用户
      x-requirement-ids: [AUTH-006]
      responses:
        '200': { $ref: '#/components/responses/UserOk' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [Users]
      operationId: updateCurrentUser
      summary: 修改当前用户设置
      x-requirement-ids: [AUTH-006, DATA-005, FX-006]
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/merge-patch+json: { schema: { $ref: '#/components/schemas/UpdateUserRequest' } }
      responses:
        '200': { $ref: '#/components/responses/UserOk' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webRefreshCookie:
      type: apiKey
      in: cookie
      name: ziji_refresh
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 认证写操作按当前用户，公开注册/密码重置按版本化匿名主体，与 API 主版本和 operationId 共同形成作用域；同键异参返回冲突。
      schema: { type: string, minLength: 16, maxLength: 100 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 资源 ETag，例如双引号包围的实体版本。
      schema: { type: string, minLength: 3, maxLength: 80 }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
...
  requestBodies:
    EmailChallengeRequest:
      required: true
      content:
        application/json: { schema: { $ref: '#/components/schemas/EmailChallengeRequest' } }
    LoginRequest:
      required: true
      content:
        application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } }
  schemas:
...
    Currency: { type: string, pattern: '^[A-Z]{3}$', examples: [CNY] }
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [requestId]
      properties: { requestId: { type: string, minLength: 1, maxLength: 100 } }
...
    EmailChallengeRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        deviceId:
          type: string
          minLength: 1
          maxLength: 200
          description: 可选防滥用信号；缺失时按来源 IP 与 MISSING_DEVICE 域标记计算设备限流摘要。
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, verificationCode, password, nickname, timezone, baseCurrency, locale]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        verificationCode: { type: string, pattern: '^[0-9]{6}$' }
        password: { type: string, minLength: 10, maxLength: 128, writeOnly: true }
        nickname: { type: string, minLength: 1, maxLength: 100 }
        timezone: { type: string, minLength: 1, maxLength: 64, default: Asia/Shanghai }
        baseCurrency: { $ref: '#/components/schemas/Currency' }
        locale: { type: string, minLength: 2, maxLength: 16 }
...
    MobileRefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 32
          maxLength: 500
          writeOnly: true
          description: 仅客户端持有一次的刷新凭据；V1 签发值为 rt1_ 加 32 字节 SecureRandom 的无填充 Base64URL，服务端不保存原文并在每次成功刷新后轮换。
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [email, challengeCode, newPassword]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        challengeCode: { type: string, pattern: '^[0-9]{6}$' }
        newPassword: { type: string, minLength: 10, maxLength: 128, writeOnly: true }
...
    UpdateUserRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nickname: { type: string, minLength: 1, maxLength: 100 }
        timezone: { type: string, minLength: 1, maxLength: 64 }
        baseCurrency: { $ref: '#/components/schemas/Currency' }
        locale: { type: string, minLength: 2, maxLength: 16 }
        amountFormat: { type: string, enum: [STANDARD, ACCOUNTING] }
...
    UserEnvelope: { $ref: '#/components/schemas/UserDataEnvelope' }
...
    MobileSessionEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [session, tokens]
          properties: { session: { $ref: '#/components/schemas/Session' }, tokens: { $ref: '#/components/schemas/TokenPair' } }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...
  responses:
    BadRequest: { description: 请求格式或字段校验失败, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Unauthenticated: { description: 未认证或会话失效, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    InvalidCredentials: { description: 邮箱不存在、密码错误、LOCKED 或 CLOSED 统一返回 INVALID_CREDENTIALS，不泄露账号存在、状态或密码 Hash 信息, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Forbidden: { description: 无对象权限或 CSRF 校验失败, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    Conflict: { description: 幂等键异参、处理中、版本冲突或唯一约束冲突；仅 code=VERSION_CONFLICT 时返回完整 versionConflict，资源不可见时按 404 返回；处理中时 Retry-After 固定为 5 秒, headers: { Retry-After: { description: 仅 IDEMPOTENCY_REQUEST_IN_PROGRESS 返回的重试等待秒数, schema: { type: integer, const: 5 } } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    RateLimited: { description: 请求过于频繁；Retry-After 为所有超限窗口中的最长剩余秒数, headers: { Retry-After: { description: 所有超限窗口中的最长剩余时间（秒）, schema: { type: integer, minimum: 1 } } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    ChallengeAccepted:
      description: 若符合发送规则则已接受
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [data, meta]
            properties:
              data:
                type: object
                additionalProperties: false
                required: [expiresIn]
                properties: { expiresIn: { type: integer, minimum: 1 } }
              meta: { $ref: '#/components/schemas/ResponseMeta' }
    UserCreated: { description: 用户已创建, headers: { ETag: { schema: { type: string } }, Location: { schema: { type: string, pattern: '^/(?!/)[^\\s]*$' } } }, content: { application/json: { schema: { $ref: '#/components/schemas/UserEnvelope' } } } }
    UserOk: { description: 用户资料, headers: { ETag: { schema: { type: string } } }, content: { application/json: { schema: { $ref: '#/components/schemas/UserEnvelope' } } } }
    InternalError: { description: 服务端无法安全完成或重放请求；不得暴露 SQL、堆栈、密码、验证码或 Token；仅已确认无业务事实的 FAILED_RETRYABLE 返回 Retry-After 5 秒, headers: { Retry-After: { description: 仅 FAILED_RETRYABLE 返回的固定重试等待秒数, schema: { type: integer, const: 5 } } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    WebSessionCreated: { description: "Web 会话已创建；响应以重复 Set-Cookie 设置 host-only 的 ziji_refresh（Secure、HttpOnly、SameSite=Strict、Path=/api/v1）与 ziji_csrf（Secure、非 HttpOnly、SameSite=Strict、Path=/api/v1），并设置 Cache-Control: no-store；刷新 Token 不出现在 JSON", content: { application/json: { schema: { $ref: '#/components/schemas/WebSessionEnvelope' } } } }
    WebSessionRefreshed: { description: "Web 会话已轮换；响应以重复 Set-Cookie 轮换同属性的 ziji_refresh 与 ziji_csrf，并设置 Cache-Control: no-store；刷新 Token 不出现在 JSON", content: { application/json: { schema: { $ref: '#/components/schemas/WebSessionEnvelope' } } } }
    MobileSessionCreated: { description: Mobile 会话与刷新 Token 已创建, content: { application/json: { schema: { $ref: '#/components/schemas/MobileSessionEnvelope' } } } }
    MobileSessionRefreshed: { description: Mobile 刷新 Token 已轮换, content: { application/json: { schema: { $ref: '#/components/schemas/MobileSessionEnvelope' } } } }
...